+++ src/me/anno/support/jvm/expression/JVMSimpleCall.kt
@@ -14,7 +14,7 @@ class JVMSimpleCall(
     val valueParameters: List<SimpleFieldExpr>,
     val enableInheritance: Boolean,
 
-    scope: Scope, origin: Int,
+    scope: Scope, origin: Long,
 ) : JVMSimpleExpr(scope, origin) {
     override fun resolveReturnType(context: ResolutionContext): Type = dst.type
 }
