+++ src/me/anno/support/jvm/expression/JVMSimpleMultiArray.kt
@@ -8,7 +8,7 @@ import me.anno.zauber.types.impl.ClassType
 
 class JVMSimpleMultiArray(
     val dst: SimpleFieldExpr, val totalType: ClassType, val dimensions: List<SimpleFieldExpr>,
-    scope: Scope, origin: Int
+    scope: Scope, origin: Long
 ) : JVMSimpleExpr(scope, origin) {
     override fun resolveReturnType(context: ResolutionContext): Type = Types.Nothing
 }
