+++ src/me/anno/support/jvm/expression/JVMSimpleThrow.kt
@@ -5,6 +5,6 @@ import me.anno.zauber.typeresolution.ResolutionContext
 import me.anno.zauber.types.Type
 import me.anno.zauber.types.Types
 
-class JVMSimpleThrow(val value: SimpleFieldExpr, scope: Scope, origin: Int) : JVMSimpleExpr(scope, origin) {
+class JVMSimpleThrow(val value: SimpleFieldExpr, scope: Scope, origin: Long) : JVMSimpleExpr(scope, origin) {
     override fun resolveReturnType(context: ResolutionContext): Type = Types.Nothing
 }
