+++ src/me/anno/support/jvm/expression/SimpleFieldExpr.kt
@@ -5,7 +5,7 @@ import me.anno.zauber.scope.Scope
 import me.anno.zauber.typeresolution.ResolutionContext
 import me.anno.zauber.types.Type
 
-class SimpleFieldExpr(val type: Type, val id: Int, scope: Scope, origin: Int) : Expression(scope, origin) {
+class SimpleFieldExpr(val type: Type, val id: Int, scope: Scope, origin: Long) : Expression(scope, origin) {
 
     override fun resolveReturnType(context: ResolutionContext): Type = type
     override fun clone(scope: Scope): Expression = this
