+++ src/me/anno/zauber/ast/rich/InnerSuperCall.kt
@@ -3,7 +3,7 @@ package me.anno.zauber.ast.rich
 class InnerSuperCall(
     val target: InnerSuperCallTarget,
     val valueParameters: List<NamedParameter>,
-    val origin: Int
+    val origin: Long
 ) {
     override fun toString(): String {
         return "$target(${valueParameters.joinToString { it.toString() }})"
