+++ src/me/anno/zauber/ast/rich/controlflow/Catch.kt
@@ -3,7 +3,7 @@ package me.anno.zauber.ast.rich.controlflow
 import me.anno.zauber.ast.rich.Parameter
 import me.anno.zauber.ast.rich.expression.Expression
 
-class Catch(val parameter: Parameter, val body: Expression, val origin: Int) {
+class Catch(val parameter: Parameter, val body: Expression, val origin: Long) {
     override fun toString(): String {
         return "catch(${parameter.name}: ${parameter.type}) { $body }"
     }
