+++ src/me/anno/zauber/ast/rich/controlflow/ContinueExpression.kt
@@ -6,7 +6,7 @@ import me.anno.zauber.scope.Scope
 import me.anno.zauber.types.Type
 import me.anno.zauber.types.Types
 
-class ContinueExpression(val label: Scope, scope: Scope, origin: Int) : Expression(scope, origin) {
+class ContinueExpression(val label: Scope, scope: Scope, origin: Long) : Expression(scope, origin) {
 
     override fun toStringImpl(depth: Int): String {
         return "continue@$label"
