+++ src/me/anno/zauber/ast/rich/controlflow/ReturnExpression.kt
@@ -6,7 +6,7 @@ import me.anno.zauber.scope.Scope
 import me.anno.zauber.types.Type
 import me.anno.zauber.types.Types
 
-class ReturnExpression(value: Expression, label: String?, scope: Scope, origin: Int) :
+class ReturnExpression(value: Expression, label: String?, scope: Scope, origin: Long) :
     ExitExpression(value, label, scope, origin) {
 
     override fun toStringImpl(depth: Int): String {
