+++ src/me/anno/zauber/ast/rich/controlflow/WhenBranchExpression.kt
@@ -10,7 +10,7 @@ class WhenCase(val condition: Expression?, val body: Expression) {
     }
 }
 
-fun whenBranchToIfElseChain(cases: List<WhenCase>, scope: Scope, origin: Int): Expression {
+fun whenBranchToIfElseChain(cases: List<WhenCase>, scope: Scope, origin: Long): Expression {
     var chain: Expression? = null
     for (i in cases.indices.reversed()) {
         val caseI = cases[i]
