+++ src/me/anno/zauber/ast/rich/expression/unresolved/CallExpression.kt
@@ -28,7 +28,7 @@ class CallExpression(
     base: Expression,
     typeParameters: List<Type>?,
     valueParameters: List<NamedParameter>,
-    origin: Int
+    origin: Long
 ) : CallExpressionBase(
     base, typeParameters,
     valueParameters, base.scope, origin
