+++ src/me/anno/zauber/ast/simple/SimpleInstruction.kt
@@ -14,6 +14,6 @@ import me.anno.zauber.scope.Scope
  *
  * todo this is effectively LLVM IR, just that names can be used multiple times
  * */
-abstract class SimpleInstruction(val scope: Scope, val origin: Int) {
+abstract class SimpleInstruction(val scope: Scope, val origin: Long) {
     abstract fun execute(): BlockReturn?
 }
